Move unstub cleanup to afterEach
